Implement a target dtype of the differentiated cube

+++ specklepy/scripts/specklepy.py
@@ -85,7 +85,8 @@ def main():
         run.inspect(files=args.files, keywords=args.keywords, save=args.save, debug=args.debug)
 
     elif args.command == 'diff':
-        differentiate_cube(files=args.files, exposure_time_prefix=args.keyword, extension=args.extension)
+        differentiate_cube(files=args.files, exposure_time_prefix=args.keyword, extension=args.extension,
+                           dtype=args.dtype)
 
     elif args.command == 'plot':
         plot = Plot.from_file(file_name=args.file, extension=args.extension, columns=args.columns, format=args.format,
